requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Email Processing
email-validator==2.1.0.post1

# Git Integration
gitpython==3.1.41

# AI & PDF Processing
openai==1.10.0
pymupdf==1.23.21

# HTTP & Webhooks
requests==2.31.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.6
